01:15

+++ src/utils/remove-file.ts
@@ -3,11 +3,23 @@ import fs from "fs/promises";
 import { errorLogger } from "@config/logger";
 import getConfig from "@config/dotenv";
 
-const removeFile = async (filePath: any) => {
-  const file = path.join(getConfig("STORAGE_UPLOAD_PDF"), filePath.filename);
-  
+const checkFileExists = async (file: string): Promise<boolean> => {
   try {
-    await fs.unlink(file);
+    await fs.access(file);
+    return true; // File eksis
+  } catch {
+    return false; // File tidak ditemukan
+  }
+}
+
+const removeFile = async (filePath: any) => {
+  const file: string = path.join(getConfig("STORAGE_UPLOAD_PDF"), filePath);
+
+  try {
+    const cek = await checkFileExists(file)
+    if (cek === true) {
+      await fs.unlink(file)
+    }
   } catch (error) {
     errorLogger.error(`ERROR REMOVE FILE: ${error}`);
   }
